Generate configurations using unity script
